refactor: consolidate tutorial integration tests
Merge tutorial-block-restriction.test.js (5 tests) into
smalruby-tutorials.test.js (3 tests) to reduce cold-start overhead.
Combined file: 8 tests in one driver instance.

Refs #240
